requirements.txt
ipdb==0.13.9
scikit-learn==1.0.2
torch==1.11.0
torchmetrics==0.10.2
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.1
transformers==4.24.0
dill==0.3.5.1
# numpy==1.18.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
unicodedata2==15.0.0
shapely==2.0.1
pandas==2.1.0
